Deteccion de tono mediante umbral y offset solamente. Interpreta el valor adecuadamente bajo condiciones ideales y con ruido. No revisa twist ni las otras cosas

+++ interpretaciontono.h
@@ -38,7 +38,7 @@ bool Umbral (float* in, float umbral);
  *
 */
 
-bool Twist (int* in);
+bool Twist (float* in);
 
 /*
  * Prueba de Offset:
